requirements.txt
asgiref
Django~=5.0.7
sqlparse
psycopg2-binary
celery~=5.4.0
redis
django-celery-beat
djangorestframework~=3.15.2
djangorestframework-simplejwt
drf-yasg
setuptools
requests~=2.32.3
django-celery-beat
python-dotenvpython-telegram-bot
python-dotenv~=1.0.1